新增鉴权功能，数据库增加字段user_role判断当前用户是否为管理员

+++ src/main/java/com/jack/usercenter/controller/UserController.java
@@ -10,7 +10,11 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpServletRequest;
+import java.util.ArrayList;
 import java.util.List;
+
+import static com.jack.usercenter.constant.UserConstant.ADMIN_ROLE;
+import static com.jack.usercenter.constant.UserConstant.USER_LOGIN_STATE;
 
 @RestController
 @RequestMapping("/user")
@@ -47,7 +51,11 @@ public class UserController {
     }
 
     @GetMapping("/search")
-    public List<User> searchUsers(String username){
+    public List<User> searchUsers(String username, HttpServletRequest request){
+        boolean flag = isAdmin(request);
+        if (!flag){
+            return new ArrayList<>();
+        }
         LambdaQueryWrapper<User> queryWrapper = new LambdaQueryWrapper<>();
         if (StringUtils.isNotBlank(username)){
             queryWrapper.like(User::getUserName, username);
@@ -56,11 +64,25 @@ public class UserController {
     }
 
     @PostMapping("/delete")
-    public boolean deleteUser(@RequestBody int id){
+    public boolean deleteUser(@RequestBody int id, HttpServletRequest request){
+        boolean flag = isAdmin(request);
+        if (!flag){
+            return false;
+        }
         if (id<=0){
             return false;
         }
         return userService.removeById(id);
+    }
+
+    /**
+     * 是否为管理员
+     * @param request
+     * @return
+     */
+    private boolean isAdmin(HttpServletRequest request){
+        User user = (User) request.getSession().getAttribute(USER_LOGIN_STATE);
+        return user != null && user.getUserRole() == ADMIN_ROLE;
     }
